destination / formatter refactoring; automatically compress triple files; made multi-threaded extraction much simpler

+++ core/src/main/scala/org/dbpedia/extraction/destinations/MarkerDestination.scala
@@ -1,26 +1,45 @@
 package org.dbpedia.extraction.destinations
 
-import java.io.{File,FileOutputStream}
+import java.io.{File,FileOutputStream,IOException}
 
-class MarkerDestination(file: File, private var start: Boolean, private var finish: Boolean)
-extends Destination {
-  
-  override def write(graph : Seq[Quad]): Unit = {
-    if (start) {
-      touch()
-      start = false
-    }
+/**
+ * Handles a marker file that signals that the extraction is either running ('start mode')
+ * or finished ('end mode').
+ * 
+ * In 'start mode', the file is created before the extraction starts (it must not already exist)
+ * and deleted after the extraction ends.
+ * 
+ * In 'end mode', the file is deleted before the extraction starts (if it already exists) 
+ * and re-created after the extraction ends.
+ * 
+ * @param file marker file
+ * @param start 'start mode' if true, 'end mode' if false. 
+ */
+class MarkerDestination(destination: Destination, file: File, start: Boolean)
+extends Destination
+{
+  override def open(): Unit = {
+    if (start) create() else delete()
+    destination.open()
   }
 
-  override def close() : Unit = {
-    if (finish) {
-      touch()
-      finish = false
-    }
+  override def write(graph: Seq[Quad]): Unit = {
+    destination.write(graph)
+  } 
+
+  override def close(): Unit = {
+    destination.close()
+    if (! start) create() else delete()
   }
   
-  private def touch(): Unit = {
-    new FileOutputStream(file).close() 
+  private def create(): Unit = {
+    if (file.exists()) throw new IOException("file '"+file+"' already exists")
+    new FileOutputStream(file).close()
+  }
+  
+  private def delete(): Unit = {
+    if (! file.exists()) return
+    if (! file.delete()) throw new IOException("failed to delete file '"+file+"'")
   }
   
 }
